Copy kernel config before building image
Some linux distros (at least Fedora 32) don't have the configuration of
the running kernel exposed at /proc/config.gz.  If this is the case,
then we somehow have to get the configuration of the host kernel into
the Docker image when building it.

* Add a script to copy the host kernel config into a file at the root of
  the repo in a (hopeully) cross-distro way
* Change the Dockerfile to copy the kernel config from the from this
  file on the host into the image
* Add an `initializeCommand` to the VS Code `.devcontainer` stuff to run
  this script automatically before building the image

There's very likely a better way to do this, but this worked for me on
Fedora 32 to get the Docker image to build automatically from VS Code.